# api/openapi.yaml
openapi: 3.0.3
info:
  title: Writers' Guild API
  version: 0.1.0
  description: |
    HTTP contract of the Writers' Guild. The Go server interface and the
    TypeScript client are generated from this file. Streaming endpoints use
    Server-Sent Events and are documented here but implemented by hand.
servers:
  - url: /
security: []
paths:
...
  /api/settings:
    get:
      operationId: getSettings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      operationId: getProject
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteProject
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
  /api/projects/{projectId}/chapters:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      operationId: listChapters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChapterSummary" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createChapter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChapterCreateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
        default: { $ref: "#/components/responses/Error" }
  /api/projects/{projectId}/bible:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      operationId: listBibleEntries
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BibleEntry" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createBibleEntry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BibleEntryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BibleEntry" }
        default: { $ref: "#/components/responses/Error" }
  /api/bible-entries/{entryId}:
    parameters:
      - $ref: "#/components/parameters/entryId"
    put:
      operationId: updateBibleEntry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BibleEntryInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BibleEntry" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteBibleEntry
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
  /api/chapters/{chapterId}:
    parameters:
      - $ref: "#/components/parameters/chapterId"
    get:
      operationId: getChapter
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updateChapter
      summary: Rename or move a chapter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChapterMetaInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteChapter
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
  /api/chapters/{chapterId}/content:
    parameters:
      - $ref: "#/components/parameters/chapterId"
    put:
      operationId: saveChapterContent
      summary: Save the chapter text; may create an autosave snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChapterContentInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChapterSaveResult" }
        "409":
          description: The chapter changed since base_hash was read
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        default: { $ref: "#/components/responses/Error" }
  /api/chapters/{chapterId}/versions:
    parameters:
      - $ref: "#/components/parameters/chapterId"
    get:
      operationId: listChapterVersions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChapterVersionSummary" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createChapterSnapshot
      summary: Take a labelled snapshot by hand
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SnapshotInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChapterVersion" }
        default: { $ref: "#/components/responses/Error" }
  /api/chapters/{chapterId}/versions/{versionId}:
    parameters:
      - $ref: "#/components/parameters/chapterId"
      - $ref: "#/components/parameters/versionId"
    get:
      operationId: getChapterVersion
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChapterVersion" }
        default: { $ref: "#/components/responses/Error" }
  /api/chapters/{chapterId}/versions/{versionId}/restore:
    parameters:
      - $ref: "#/components/parameters/chapterId"
      - $ref: "#/components/parameters/versionId"
    post:
      operationId: restoreChapterVersion
      summary: Snapshot the current text, then replace it with the version
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    chapterId:
      name: chapterId
      in: path
      required: true
      schema: { type: string, format: uuid }
    versionId:
      name: versionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    entryId:
      name: entryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    Settings:
      type: object
      required: [scene_token_limit, autosave_snapshot_minutes]
      properties:
        scene_token_limit: { type: integer }
        autosave_snapshot_minutes: { type: integer }
    SettingsInput:
      type: object
      required: [scene_token_limit, autosave_snapshot_minutes]
      properties:
        scene_token_limit: { type: integer, minimum: 500, maximum: 200000 }
        autosave_snapshot_minutes: { type: integer, minimum: 1, maximum: 1440 }
...
    Project:
      type: object
      required: [id, name, description, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 5000 }
    ChapterSummary:
      type: object
      required: [id, project_id, title, position, content_hash, content_length, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        title: { type: string }
        position: { type: integer }
        content_hash: { type: string }
        content_length: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Chapter:
      type: object
      required: [id, project_id, title, position, content_md, content_hash, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        title: { type: string }
        position: { type: integer }
        content_md: { type: string }
        content_hash: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ChapterCreateInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 300 }
        content_md: { type: string }
    ChapterMetaInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 300 }
        position: { type: integer }
    ChapterContentInput:
      type: object
      required: [content_md]
      properties:
        content_md: { type: string }
        base_hash:
          type: string
          description: Hash the client last loaded; when set and stale the save is refused with 409
    ChapterSaveResult:
      type: object
      required: [chapter, snapshot_created]
      properties:
        chapter: { $ref: "#/components/schemas/Chapter" }
        snapshot_created: { type: boolean }
    ChapterVersionSummary:
      type: object
      required: [id, chapter_id, kind, label, content_hash, content_length, created_at]
      properties:
        id: { type: string, format: uuid }
        chapter_id: { type: string, format: uuid }
        kind: { type: string, enum: [autosave, manual, pre_revision, pre_restore] }
        label: { type: string }
        content_hash: { type: string }
        content_length: { type: integer }
        created_at: { type: string, format: date-time }
    ChapterVersion:
      type: object
      required: [id, chapter_id, kind, label, content_md, content_hash, created_at]
      properties:
        id: { type: string, format: uuid }
        chapter_id: { type: string, format: uuid }
        kind: { type: string, enum: [autosave, manual, pre_revision, pre_restore] }
        label: { type: string }
        content_md: { type: string }
        content_hash: { type: string }
        created_at: { type: string, format: date-time }
    SnapshotInput:
      type: object
      properties:
        label: { type: string, maxLength: 200 }
    BibleSection:
      type: string
      enum: [premise, character, setting, timeline, style, chapter_summary]
    BibleEntry:
      type: object
      required: [id, project_id, section, title, fields, position, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        section: { $ref: "#/components/schemas/BibleSection" }
        title: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
        chapter_id: { type: string, format: uuid }
        position: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BibleEntryInput:
      type: object
      required: [section, title, fields]
      properties:
        section: { $ref: "#/components/schemas/BibleSection" }
        title: { type: string, maxLength: 300 }
        fields:
          type: object
          additionalProperties: { type: string }
        chapter_id: { type: string, format: uuid }
        position: { type: integer }
...