# internal/db/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"time"

	"github.com/google/uuid"
)

type BibleEntry struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	ProjectID uuid.UUID
	Section   string
	Title     string
	Fields    []byte
	ChapterID uuid.NullUUID
	Position  int32
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Chapter struct {
	ID          uuid.UUID
	UserID      uuid.UUID
	ProjectID   uuid.UUID
	Title       string
	Position    int32
	ContentMd   string
	ContentHash string
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type ChapterVersion struct {
	ID          uuid.UUID
	UserID      uuid.UUID
	ChapterID   uuid.UUID
	Kind        string
	Label       string
	ContentMd   string
	ContentHash string
	CreatedAt   time.Time
}

type ModelCall struct {
	ID               uuid.UUID
	UserID           uuid.UUID
	RunID            uuid.NullUUID
	WriterID         uuid.NullUUID
	GenerationName   string
	ModelAlias       string
	PromptTokens     int32
	CompletionTokens int32
	CostUsd          float64
	CostEstimated    bool
	LatencyMs        int32
	Status           string
	Error            string
	CreatedAt        time.Time
}

type Project struct {
	ID          uuid.UUID
	UserID      uuid.UUID
	Name        string
	Description string
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type Run struct {
	ID               uuid.UUID
	UserID           uuid.UUID
	ProjectID        uuid.NullUUID
	ChapterID        uuid.NullUUID
	Kind             string
	Status           string
	TraceID          string
	Params           []byte
	Result           []byte
	Error            string
	CostUsd          float64
	CostEstimated    bool
	PromptTokens     int32
	CompletionTokens int32
	CreatedAt        time.Time
	StartedAt        *time.Time
	FinishedAt       *time.Time
}

type RunEvent struct {
	RunID     uuid.UUID
	Seq       int32
	Type      string
	Payload   []byte
	CreatedAt time.Time
}

type User struct {
	ID           uuid.UUID
	Username     string
	DisplayName  string
	PasswordHash string
	Role         string
	DisabledAt   *time.Time
	AuthVersion  int32
	CreatedAt    time.Time
	UpdatedAt    time.Time
}

type UserSetting struct {
	UserID                  uuid.UUID
	SceneTokenLimit         int32
	AutosaveSnapshotMinutes int32
	UpdatedAt               time.Time
}

type Writer struct {
	ID           uuid.UUID
	UserID       uuid.UUID
	Name         string
	Slug         string
	ModelAlias   string
	SystemPrompt string
	Roles        []string
	Enabled      bool
	Temperature  float64
	IsSystem     bool
	CreatedAt    time.Time
	UpdatedAt    time.Time
}